+++ CHANGES.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased] - ReleaseDate
 
+### Added
+
+- New `xot.level_order()` method to traverse the tree in level order.
+
 ## [0.13.0] - 2023-02-23
 
 ### Added
@@ -331,8 +335,8 @@
 Initial public release.
 
 <!-- next-url -->
-[Unreleased]: https://github.com/faassen/xot/compare/v0.13.0...HEAD
 
+[unreleased]: https://github.com/faassen/xot/compare/v0.13.0...HEAD
 [unreleased]: https://github.com/faassen/xot/compare/v0.12.1...v0.13.0
 [unreleased]: https://github.com/faassen/xot/compare/v0.12.0...v0.12.1
 [unreleased]: https://github.com/faassen/xot/compare/v0.11.8...v0.12.0